chore: move Integerlist to db-api
