+++ crates/storage/db/src/tables/codecs/fuzz/mod.rs
@@ -17,9 +17,6 @@ macro_rules! impl_fuzzer_with_input {
                 use reth_db_api::table;
 
                 #[allow(unused_imports)]
-
-
-                #[allow(unused_imports)]
                 use reth_primitives_traits::*;
 
                 #[allow(unused_imports)]
